updating smart contract folders and fixing bugs.

+++ app/NFTs/scripts/end_royalty_contract.py
@@ -1,18 +1,35 @@
 import NFTs.scripts.transaction as trx
+from os.path import isdir, isfile
+
 
 def buy_token_with_royalty(tmp, wallet_skey_path, wallet_addr, script_addr, cost, royalty, datum_hash, plutus_script, seller_addr, royalty_addr, policy_id, token_name, collateral, flag):
     print('\nSTART OF BUYING ROYALTY SALE CONTRACT')
     print(policy_id)
     print(token_name)
+
+    # Ensure the tmp folder exists
+    if isdir(tmp) is False:
+        print('The directory:', tmp, 'does not exist.')
+        return False
     
+    # Clean start each time
     trx.delete_contents(tmp)
     trx.protocol(tmp, flag)
     
     trx.utxo(wallet_addr, tmp, 'utxo.json', flag)
+    # Check if wallet address is correct
+    if isfile(tmp+'utxo.json') is False:
+        print('The file:', tmp+'utxo.json', 'does not exist.')
+        return False
+    
     utxo_in, utxo_col, currencies, collat_flag, _ = trx.txin(tmp, 'utxo.json', collateral)
     if collat_flag is True:
         #tmp
         trx.utxo(script_addr, tmp, 'utxo_script.json', flag)
+        if isfile(tmp+'utxo_script.json') is False:
+            print('The file:', tmp+'utxo_script.json', 'does not exists')
+            return False
+        
         _, _, script_currencies, _, data_list = trx.txin(tmp, 'utxo_script.json')
         contract_utxo_in = utxo_in
         for key in data_list:
@@ -33,7 +50,16 @@ def buy_token_with_royalty(tmp, wallet_skey_path, wallet_addr, script_addr, cost
         ]
         # print('BLOCK:', block)
         # print(utxo_out)
-        trx.build(tmp, wallet_addr, final_tip, contract_utxo_in, utxo_col, utxo_out, additional_data, flag)
+        check_status = trx.build(tmp, wallet_addr, final_tip, contract_utxo_in, utxo_col, utxo_out, additional_data, flag)
+        if check_status != 0:
+            print('The transaction build has failed.')
+            return False
+        
+        # Ensure the tmp folder exists
+        if isfile(wallet_skey_path) is False:
+            print('The file:', wallet_skey_path, 'does not exist.')
+            return False
+        
         signers = [
             '--signing-key-file',
             wallet_skey_path,
@@ -41,6 +67,8 @@ def buy_token_with_royalty(tmp, wallet_skey_path, wallet_addr, script_addr, cost
         trx.sign(tmp, signers, flag)
         trx.submit(tmp, flag)
         print('\nEND OF BUYING ROYALTY SALE CONTRACT\n')
+        return True
 
     else:
-        print("SPLIT UP UTXO")
+        print("The wallet did not account for collateral.")
+        return False
